Code review changes and moved database initialization helper functions to db.ts to remove circular dependencies
