+++ src/bot.ts
@@ -3,7 +3,6 @@ dotenv.config();
 
 import Discord from 'discord.js';
 import yaml from 'js-yaml';
-import fs from 'fs';
 import Commando from 'discord.js-commando';
 import path from 'path';
 
@@ -13,15 +12,14 @@ import { messageListener } from './components/messageListener';
 import { initEmojis } from './components/emojis';
 import { createSuggestionCron, createBonusInterviewerListCron } from './components/cron';
 import { readFileSync } from 'fs';
+import { vars } from './constants';
 
-const ENV: string = process.env.NODE_ENV || '.';
-const vars = JSON.parse(readFileSync(`./config/${ENV}/vars.json`, 'utf-8'));
 const NOTIF_CHANNEL_ID: string = vars.NOTIF_CHANNEL_ID;
 const BOT_TOKEN: string = process.env.BOT_TOKEN || '.';
 const BOT_PREFIX = '.';
 
 // initialize Commando client
-const botOwners = yaml.load(fs.readFileSync('config/owners.yml', 'utf8')) as string[];
+const botOwners = yaml.load(readFileSync('config/owners.yml', 'utf8')) as string[];
 export const client = new Commando.Client({
   owner: botOwners,
   commandPrefix: BOT_PREFIX,
